[kitsune] Simplify GPU-centric tapir targets
CudaABI and HipABI had custom argument orders for the outlined functions
consisting of the bodies of tapir loops. This made them more complicated than
necessary since nothing else in the code was particularly dependent on the
order of the arguments. This has been fixed so those two tapir targets just
use the implementation of setupLoopOutlineArgs from the default
LoopOutlineProcessor for most of the work and add a small amount of additional
code. Some code was also cleaned up to use the newer API in llvm::Function to
get the nth llvm::Argument. In addition to updating the tests to reflect these
changes, some were also tweaked to avoid very long lines and to make it easier
to determine what exactly was being matched.