[kitsune] Add print-before-first command line option
Since Tapir currently requires optimizations, it is difficult to get the IR as
soon as it is generated by the frontend. Using -S -emit-llvm will only produce
IR that has been through the entire optimization pipeline. This commit adds a
-print-before-first option to both clang and flang. This prints the IR to stderr
just before the optimization passes are run. This is useful when debugging the
frontend to see exactly what IR is being produced.